+++ src/0_session.js
@@ -10,9 +10,10 @@ goog.provide('session');
  * @param {BranchStorage} storage
  * @return {Object}
  */
-session.get = function(storage) {
+session.get = function(storage, first) {
+	var sessionString = first ? 'branch_session_first' : 'branch_session';
 	try {
-		return goog.json.parse(storage.get('branch_session')) || null;
+		return goog.json.parse(storage.get(sessionString)) || null;
 	}
 	catch (e) {
 		return null;
@@ -23,6 +24,7 @@ session.get = function(storage) {
  * @param {BranchStorage} storage
  * @param {Object} data
  */
-session.set = function(storage, data) {
+session.set = function(storage, data, first) {
 	storage.set('branch_session', goog.json.serialize(data));
+	if (first) { storage.set('branch_session_first', goog.json.serialize(data)); }
 }