+++ test/1_api.js
@@ -130,6 +130,7 @@ describe('Server', function() {
 	describe('Resources', function() {
 		describe('/v1/open', function() {
 			it('should pass in branch_key and browser_fingerprint_id', function(done) {
+				storage['set']('use_jsonp', false, "session");
 				var assert = testUtils.plan(5, done);
 				server.request(resources.open, testUtils.params({ "is_referrable": 1 }), storage, assert.done);
 				assert.equal(requests.length, 1, 'Request made');
@@ -171,6 +172,7 @@ describe('Server', function() {
 			});
 
 			it('should pass without branch_key but with app_id', function(done) {
+				storage['set']('use_jsonp', false, "session");
 				var assert = testUtils.plan(2, done);
 				server.request(resources.open, testUtils.params({ "app_id": "5680621892404085", "is_referrable": 1 }, [ 'branch_key' ]), storage, assert.done);
 				assert.equal(requests.length, 1, 'Request made');
@@ -213,6 +215,7 @@ describe('Server', function() {
 
 		describe('/v1/profile', function() {
 			it('should pass in branch_key and identity', function(done) {
+				storage['set']('use_jsonp', false, "session");
 				var assert = testUtils.plan(5, done);
 				server.request(resources.profile, testUtils.params({ "identity": "test_id" }), storage, assert.done);
 
@@ -265,6 +268,7 @@ describe('Server', function() {
 
 		describe('/v1/logout', function() {
 			it('should pass in branch_key and session_id', function(done) {
+				storage['set']('use_jsonp', false, "session");
 				var assert = testUtils.plan(5, done);
 				server.request(resources.logout, testUtils.params({ }), storage, assert.done);
 
@@ -308,6 +312,7 @@ describe('Server', function() {
 
 		describe('/v1/referrals', function() {
 			it('should pass in identity_id', function(done) {
+				storage['set']('use_jsonp', false, "session");
 				var assert = testUtils.plan(4, done);
 				server.request(resources.referrals, testUtils.params({ }), storage, assert.done);
 
@@ -338,6 +343,7 @@ describe('Server', function() {
 
 		describe('/v1/credits', function() {
 			it('should pass in identity_id', function(done) {
+				storage['set']('use_jsonp', false, "session");
 				var assert = testUtils.plan(4, done);
 				server.request(resources.credits, testUtils.params({ }), storage, assert.done);
 
@@ -388,6 +394,7 @@ describe('Server', function() {
 
 		describe('/v1/redeem', function() {
 			it('should pass in branch_key, identity_id, amount, and bucket', function(done) {
+				storage['set']('use_jsonp', false, "session");
 				var assert = testUtils.plan(5, done);
 				server.request(resources.redeem, testUtils.params({ "amount": 1, "bucket": "testbucket" }), storage, assert.done);
 
@@ -450,6 +457,7 @@ describe('Server', function() {
 
 		describe('/v1/link', function() {
 			it('should pass in branch_key and identity_id', function(done) {
+				storage['set']('use_jsonp', false, "session");
 				var assert = testUtils.plan(5, done);
 
 				server.request(resources.link, testUtils.params(), storage, assert.done);
@@ -500,6 +508,7 @@ describe('Server', function() {
 
 		describe('/l', function() {
 			it('should pass in link_url and click', function(done) {
+				storage['set']('use_jsonp', false, "session");
 				var assert = testUtils.plan(4, done);
 				server.request(resources.linkClick, testUtils.params({ "link_url": "3hpH54U-58", "click": "click" }), storage, assert.done);
 
@@ -545,6 +554,7 @@ describe('Server', function() {
 			var metadataString = "&metadata.url=testurl&metadata.user_agent=test_agent&metadata.language=test_language";
 
 			it('should pass in branch_key, session_id, event and metadata', function(done) {
+				storage['set']('use_jsonp', false, "session");
 				var assert = testUtils.plan(5, done);
 
 				server.request(resources.event, testUtils.params({ "event": "testevent", "metadata": metadata }), storage, assert.done);
@@ -614,6 +624,7 @@ describe('Server', function() {
 
 		describe('/v1/creditHistory', function() {
 			it('should pass in branch_key and session_id', function(done) {
+				storage['set']('use_jsonp', false, "session");
 				var assert = testUtils.plan(4, done);
 
 				server.request(resources.creditHistory, testUtils.params(), storage, assert.done);